+++ CinemaBox.Model/Imdb/Movie/MovieModelScrapping.cs
@@ -19,7 +19,7 @@ public class MovieModelScrapping
     public long? VoteCount { get; set; }
     public long? Winner { get; set; }
     public long? Nomination { get; set; }
-    public byte? TopRank { get; set; }
+    public int? TopRank { get; set; }
     public byte? OscarNominations { get; set; }
     public byte? OscarWins { get; set; }
     public List<string>? Genres { get; set; }
